Move icon-library to internal
