+++ packages/components/src/index.ts
@@ -6,7 +6,6 @@ export * from './FocusTrap';
 export * from './hooks/useUniqueId';
 export * from './hooks/useWindowDimensions';
 export * from './Icon';
-export type { IconId } from './icon-library';
 export * from './InfoBox';
 export * from './injectStyle';
 export { Interactive } from './Interactive';
